chore(meta): move dependency method to MetaLeader
